requirements.txt
# Core Web Framework
Flask==3.1.0
Werkzeug==3.1.3
asgiref==3.8.1

# Image Processing
Pillow==11.0.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
numpy==1.26.4
scikit-image==0.25.0

# OCR Engines
pytesseract==0.3.13
easyocr==1.7.2
paddlepaddle==2.6.2
paddleocr==2.9.1
surya-ocr==0.8.0

# Text Processing
python-Levenshtein==0.26.1
fuzzywuzzy==0.18.0
regex==2024.11.6

# Deep Learning
torch==2.5.1
torchvision==0.20.1
tensorflow-intel==2.18.0
keras==3.7.0

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
colorlog==6.9.0
filetype==1.2.0

# Optional GPU Support
# cuda-python>=11.8.0
# cupy>=12.0.0

# Notes:
# 1. Tesseract-OCR must be installed separately:
#    Windows: https://github.com/UB-Mannheim/tesseract/wiki
#    Don't forget to add Turkish language data during installation
# 2. Make sure Tesseract is in your system PATH
